Add test for null value and indicator of variable-length data (SQL Server).
Set separate value/indicator buffers for all SQLBindCol,
not only for non-BLOB data.
This is an attmpt to *evaluate* if it is a correct fix for discovered
issue with SQLFetch does not setting null indicator for nullable
variable-length column.

See the code inline comment for explanation of the observed problem.
See also comments of this PR #166 for detailed discussion.

Related to PR #165.